#18 fix: useeffect and moved to context
