+++ README.md
@@ -1,6 +1,6 @@
 # C I N E C I T Y
 
-Cinecity is a web application that allows you to view, sort and filter a film catalog using the API from [The Movie Database](https://developer.themoviedb.org/docs). The design of the interface, search and ordering is inspired by websites focused on author cine programming and art cine such as the page of the [Cineteca Nacional](https://www.cinetecanacional.net/#gsc. tab=0) y [Ambulante](https://www.ambulante.org/).
+Cinecity is a web application that allows you to view, sort and filter a film catalog using the API from [The Movie Database](https://developer.themoviedb.org/docs). The design of the interface, search and ordering is inspired by websites focused on author cine programming and art cine such as the page of the [Cineteca Nacional](https://www.cinetecanacional.net/#gsc.tab=0) y [Ambulante](https://www.ambulante.org/).
 
 ## Table of contents
 
@@ -37,13 +37,18 @@ It is possible to navigate the content of the catalog using the minimalist page 
 
 ## 2. Technology Stack
 
-_Cinecity_ fue desarrollado utilizando las siguientes tecnologías. 
+_Cinecity_  was developed using the following technologies.
 
 ### React
+React was the core framework used to structure and build the frontend of _Cinecity_. It facilitated the creation of custom components, services, and modules that formed the user interface. React's routing system was used to manage navigation between different parts of the application, enabling seamless transitions and user interactions.
 
 ### TypeScript
 
+TypeScript was the primary programming language used to write the frontend code. It added strong typing to the application, catching errors during development and providing better code quality and maintainability. It was instrumental in defining the data models, logic, and interactions within _Cinecity_.
+
 ### CSS
+
+ CSS was employed to style the user interface of _Cinecity_. 
 
 ## 3. Design and prototype
