+++ cinecity/src/FilterContext.tsx
@@ -1,5 +1,7 @@
 import React, { createContext, useState, useContext, Dispatch, SetStateAction } from 'react';
 import { Movie, Genre } from './services/types';
+import { useEffect } from 'react';
+import { buildApiUrl, API_KEY } from './services/movie-request';
 
 type FilterContextType = {
     movies: Movie[];
@@ -52,14 +54,14 @@ export const FilterContextProvider = ({ children }: FilterContextProviderProps) 
         urlApi: 'https://api.themoviedb.org/3/discover/movie?api_key=0a0e1e1b0b0c0d0e0f0a0b0c0d0e0f0a&language=en-US',
     });
 
-    const setMovies = (movies: Movie[]) => { 
+    const setMovies = (movies: Movie[]) => {
         setAppState((prevState) => ({
             ...prevState,
             movies: movies
         }));
     }
-    
-    const setSortMovies = (order: string) => { 
+
+    const setSortMovies = (order: string) => {
         setAppState((prevState) => ({
             ...prevState,
             sortBy: order
@@ -91,11 +93,29 @@ export const FilterContextProvider = ({ children }: FilterContextProviderProps) 
         setAppState((prevState) => ({
             ...prevState,
             urlApi: url
-        }));    
+        }));
     }
+    //Revisar libreria react query
+    useEffect(() => {
+        const apiUrl = buildApiUrl(appState.page, appState.genre, appState.sortBy);
+
+        const options = {
+            method: 'GET',
+            headers: {
+                accept: 'application/json',
+                Authorization: API_KEY,
+            }
+        };
+        fetch(apiUrl, options)
+            .then((response) => response.json())
+            .then((data) => {
+                setMovies(data.results)
+            })
+            .catch((error) => console.error(error));
+    }, [appState.sortBy, appState.genre, appState.page]);
 
     return (
-        <FilterContext.Provider value={{ appState, setAppState, setMovies,setCurrentPage, setSortMovies, setSelectedGenre, setSelectedSort, setUrlApi }}>
+        <FilterContext.Provider value={{ appState, setAppState, setMovies, setCurrentPage, setSortMovies, setSelectedGenre, setSelectedSort, setUrlApi }}>
             {children}
         </FilterContext.Provider>
     );
